bug: int 로 인식되어 crash 발생하는 내용 수정

+++ util-compose/src/main/kotlin/team/duckie/app/android/util/compose/constants.kt
@@ -8,4 +8,4 @@
 package team.duckie.app.android.util.compose
 
 /** 상수 모음 */
-const val HEIGHT_RATIO_FULL_SCREEN_HORIZONTAL_16 = 240 / 328
+const val HEIGHT_RATIO_FULL_SCREEN_HORIZONTAL_16 = 240.toFloat() / 328
